Things I've worked on so far
dice.py contains the dice class which should be used any time the dice
is rolled, and returns with the actual dice numbers to be scored.
scoring.py contains all functions that are used to calculate score
choices. As of now yahtzee_txt.py is the main, and includes the Card
class, which will keep track of players scores. test.py is for messing
with stuff in other files.