Two player works
Set card and round as instance variables (that took me longer than it
should have to figure out) and created a round function that keeps
global variables out.

Now that it all works as intended, I am going to focus on learning GUI
development, because text yahtzee is kind of lame.